+++ eva/src/components/dashBoardInside/dashConstructorScheme/dashConstructorSchemes.vue
@@ -8,7 +8,12 @@
       margin: '0 10px',
     }"
   >
-    <div class="dash-constructor-schemes__options">
+    <div
+      class="dash-constructor-schemes__options"
+      :class="{
+        'dash-constructor-schemes__options--is-keymap-open': isKeymapOpen,
+      }"
+    >
       <v-tooltip
         bottom
         :color="theme.$accent_ui_color"
@@ -137,14 +142,20 @@
     </div>
     <!--Keymap-panel-->
     <dash-constructor-schemes-keymap
+      ref="keymap"
       v-model="isKeymapOpen"
+      @changeKeymapTab="setPanelBottomOffset"
     />
     <!--Drag-and-drop panel-->
     <div
       ref="dndPanelContainer"
       class="dash-constructor-schemes__dnd-panel-container"
+      :style="{
+        'bottom': `${panelBottomOffset}px`,
+      }"
       :class="{
         'dash-constructor-schemes__dnd-panel-container--active': dndPanel,
+        // 'dash-constructor-schemes__dnd-panel-container--is-keymap-open': isKeymapOpen
       }"
     >
       <div class="row justify-end">
@@ -182,7 +193,7 @@
                       Настройки:
                     </v-expansion-panel-header>
                     <v-expansion-panel-content>
-                      <div class="dash-constructor-schemes__options">
+                      <div class="dash-constructor-schemes__inner-options">
                         <!--TODO: Возможно стоит вынести в отдельный компонент-->
                         <div class="row">
                           <div class="col-12">
@@ -817,11 +828,18 @@ export default {
       selectedDataType: '',
       dataSelectedNode: null,
       isKeymapOpen: false,
+      panelBottomOffset: 10,
     };
   },
   computed: {
     dashFromStore() {
       return this.$store.state[this.idDashFrom];
+    },
+    primitivesFromStore() {
+      if (this.dashFromStore[this.idFrom].options.primitivesLibrary) {
+        return JSON.parse(this.dashFromStore[this.idFrom].options.primitivesLibrary);
+      }
+      return [];
     },
     savedGraph() {
       return this.dashFromStore.savedGraph || '';
@@ -844,6 +862,9 @@ export default {
           this.constructorSchemes.updateDataInNode(value);
         }
       },
+    },
+    isKeymapOpen() {
+      this.setPanelBottomOffset();
     },
   },
   mounted() {
@@ -870,7 +891,7 @@ export default {
         dndPanelElem: this.$refs.dndPanel,
         elem: this.$refs.graphComponent,
         dataRest: this.mockData,
-        iconsList: this.iconsList,
+        iconsList: this.primitivesFromStore,
         elementDefaultStyles: this.elementDefaultStyles,
         openDataPanelCallback: this.openDataPanel,
         closeDataPanelCallback: this.closeDataPanel,
@@ -946,6 +967,11 @@ export default {
     openKeymapPanel() {
       this.isKeymapOpen = true;
     },
+    setPanelBottomOffset() {
+      this.$nextTick().then(() => {
+        this.panelBottomOffset = this.isKeymapOpen ? this.$refs.keymap.$el.clientHeight + 5 : 10;
+      });
+    },
   },
 };
 </script>
@@ -960,7 +986,11 @@ export default {
     top: 0;
     display: flex;
     align-items: center;
-    z-index: 1;
+    z-index: 10;
+  }
+  &__inner-options {
+    display: flex;
+    align-items: center;
   }
   &__keymap-button {
     position: absolute;
@@ -997,7 +1027,7 @@ export default {
   }
   &__dnd-panel-container, &__data-panel {
     color: var(--main_text);
-    z-index: 1;
+    z-index: 10;
     position: absolute;
     top: 5px;
     bottom: 15px;
@@ -1013,10 +1043,14 @@ export default {
   &__dnd-panel-container {
     left: 0;
     border-radius: 0 4px 4px 0;
+    transition: all .2s ease;
     transform: translateX(-100%);
     &--active {
       transform: translateX(0);
       pointer-events: all;
+    }
+    &--is-keymap-open {
+      bottom: 405px;
     }
   }
   &__data-panel {