+++ eva/src/js/componentsSettings.js
@@ -261,7 +261,7 @@ export default {
     csvg: ['tooltip', 'backgroundImage'],
     ygraph: [],
     bush: [],
-    map: ['osmserver', 'primitivesLibrary'],
+    map: ['osmserver', 'primitives', 'primitivesLibrary'],
     heatmap: [
       'dataFormat',
       'x',
@@ -300,7 +300,7 @@ export default {
       'metricGroup',
       'scatterPlotLegend',
     ],
-    constructorSchemes: ['visible', 'level', 'pinned'],
+    constructorSchemes: ['visible', 'level', 'pinned', 'primitives', 'primitivesLibrary'],
   },
   optionFields: [
     // описание типов полей и их характеристик
@@ -668,7 +668,6 @@ export default {
         return this.$store.state[this.idDash][this.element]?.availableTableTitles;
       },
     },
-
     {
       option: 'columnCount',
       description: 'Количество столбцов',
@@ -827,6 +826,18 @@ export default {
       elem: 'switch',
     },
 
+    // fullWidthGroup
+    {
+      group: 'Библиотека примитивов',
+      option: 'primitives',
+    },
+    {
+      optionGroup: 'primitives',
+      option: 'primitivesLibrary',
+      elem: 'code-editor',
+      isFullWidth: true,
+    },
+
   ],
   reporstElements: [
     'table',
@@ -908,6 +919,7 @@ export default {
     fromDataSearches: [
       'menu',
       'picker',
+      'constructorSchemes',
     ],
   },
 };